+++ src/releases/experimental.rs
@@ -6,19 +6,19 @@ use serde::{Deserialize, Serialize};
 use std::path::PathBuf;
 
 #[derive(Clone, Debug, Default, Deref, DerefMut, Deserialize, PartialEq, Serialize)]
-pub struct Branched(Vec<Package>);
+pub struct Experimental(Vec<Package>);
 
 #[async_trait]
-impl ReleaseType for Branched {
+impl ReleaseType for Experimental {
     async fn fetch() -> Self {
-        Self::fetch_from_builder(BuilderBuildsType::Branched).await
+        Self::fetch_from_builder(BuilderBuildsType::Experimental).await
     }
 
     fn get_name(&self) -> String {
-        String::from("branched")
+        String::from("experimental")
     }
 
     fn get_db_path(&self) -> PathBuf {
-        get_setting().databases_dir.join("branched.bin")
+        get_setting().databases_dir.join("experimental.bin")
     }
 }